Observed RX queue size from symbols used and pointer behavior
Apparently the RX queue is 6 packets of up to 1604 bytes.
With 36 bytes (EP_OFFSET) in front of a packet + 14 bytes for the
reconstructed ethernet header, if we subtract those 50 bytes,
we get a maximum payload length of 1554 bytes.

While the code comment says the MAC header is 32bytes, that only
seems to be true for 802.11abg dn for 802.11n it is 36bytes.
Together with the WPA2+CCMP 16 byte encryption header this
would then exactly match the presumed 50 bytes.

Also it looks like the maximum MTU for 802.11 without headers
is 2304 bytes, but there pretty much isn't enough space for
that in this setup.